+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-## 1.0.0.0 -- September 2024
+## 1.1.0.0 -- October 2024
+
+* Added `Zipper` modules for non-empty radix trees;
+
+* Removed `Cursor`s from radix tree modules.
+  Their functionality is subsumed by the appropriate `Zipper` modules.
+
+* Added `Pointer` modules for strict radix trees;
+
+## 1.0.0.2 -- September 2024
 
 * Fixed `Data.Zebra.Word.fillRange`.
   Previously it produced malformed trees in certain cases.